pyproject.toml
[project]
name = "synapse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "George Zhang", email = "georgezhangtj97@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.28.1",
  "pydantic-ai-slim>=0.1.3",
  "logfire[httpx]>=3.14.0",
  "rich>=14.0.0",
  "trio>=0.30.0",
  "pyyaml>=6.0.2",
  "pydantic-settings>=2.9.1",
]

[dependency-groups]
dev = [
  "ruff>=0.3.4",
  "pyright>=1.1.349"
]

[tool.ruff]
line-length = 120
target-version = "py313"
include = ["src/synapse/**/*.py"]

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportUnnecessaryTypeIgnoreComment = true
include = ["src/synapse/**/*.py"]
venvPath = ".venv"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint]
extend-select = ["Q", "RUF100", "C90", "UP", "I"]
flake8-quotes = { inline-quotes = "single", multiline-quotes = "double" }
isort = { combine-as-imports = true }
mccabe = { max-complexity = 15 }

[project.scripts]
synapse = "synapse.main:run_main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/synapse"]
